feat: 重构 MCPConfig 组件，提取工具项为独立组件 MCPToolItem，优化参数展示逻辑，增强代码可读性
